feat: Open mod config files in the app
added more info sent to the front end to determin to show an
additional button for editing a mods config files

the front end will send a Vec<SharedString> to rust for all ini files
that are registred with a mod. We will loop over them and spawn a thread
for each to open the file in notepad. These are stored as full length
paths apon deserialization of RegMod's.