parser now checks if saved files are .ini
Parser now splits read files from the ini into
    files: of type non-ini
    config_files: files ending in .ini

This allows the deserializer to not have to do this and allows
the modification of toggle_files() to toggle all registered files
except configuration files.

This allows the user to open and edit a configuration file when a
mod is in the disabled state.