updated deserializer
The backend RegMod.name now is stored the same way as it is saved.
The deserializer now takes care of formating it to the user.
This way we do need to format the key in main more often but
RegMod now plays nicer with r/w function and toggle_files()