Feat: Auto import installed mods
On first startup you have the selection to auto scan for installed mods
this will give them names based off the file name, it is able to import
mods that are in the .disabled state. It will include any files
in a directory with the same name located adjacently.

It does not take into concideration if multiple files in ./mods/
are supposed to be grouped together