Optimized out cfg.collect_mods()
replaced file_registered() with cfg.files() this way we can only parse
out the data we need to get this check done, and this way we can
optimize the type to make the time complexity of checking all files much
faster. we now use a HashSet for quick lookups

also added a cfg.keys() so we can eliminate another call to
cfg.collect_mods() this one makes sure all the keys in the ini is valid
and if not will correct the key entries by calling sync keys. this
also returns the keys stored in a HashSet for quick lookups

the static RESTRICTED_FILES, also was moved over to use a HashSet