Optimization: remove call to collect_mods
also removed a call to deserialize all. we were able to remove these
by creating a method on cfg that allows us to parse out a singular
`RegMod` from our ini data. in the case of `add_to_mod` we only
deserialize the updated split files and then update the front end data

we are now able to deserialize just the individual parts of our `RegMod`
now that the deserialize_all process is abstracted into 3 function calls

also fixed a bug that we previously were not running into with
`IniProperty::<PathBuf>::read()` now we force the caller to supply
a prefix and let them know if they messed up the call structure by
returning an error