move private ui fn to global scope
this is handy so we can easily call this from the native rust code
if we ever update order information from rust then we can re-draw the
ui elements that display this information properly.

This is done because of how slint does not currently allow two-way
bindings with data stored in structs or arrays (our case uses an array
of a stuct we define). Currently the two-way binding breaks as soon as
the user interacts with the UI element